Add rpm -a support to rpmls
And grow a more robust handling of options and arguments. Using the -p
option should be based on whether the argument refers to a file instead
of a heuristic based on a glob expression `*.[sr]pm` that somehow will
accept `.spm` as a valid RPM extension.

It becomes possible to compose commands like this:

    rpmls -la varnish\* |
    awk '/^d/ {print $NF}' |
    sort