Replace spectool with rpmdev-spectool
The old spectool was a weird Perl program that was difficult to debug
and maintain, especially as RPM spec files became more and more complex.

This new implementation leverages the RPM Python bindings and should
be considerably easier to maintain and extend.

Thanks to Fabio Valentini for contributing this new implementation!