bumpspec, newspec: Add ability to use legacy datestamp when desirable
If a spec file is being created on a system with an old RPM version
that does not support the richer datestamp format, we should not use
full datestamps when creating the spec.

Additionally, if someone wants to bump the spec file with the legacy
datestamp format that is still supported (for a spec that supports
targeting older systems using versions of RPM that do not support it),
then the new --legacy-datestamp switch for bumpspec will generate
compatible changelog entries for it.